+++ K-Content/src/main/resources/static/css/admin-body-header.css
@@ -23,7 +23,7 @@ header div.logo img {
     width: 160px;
     height: 90px;
     margin: 0;
-    padding: 0x
+    padding: 0px;
 }
 
 header ul.icons {
